examples/dev/EMG/gmViewerAcc/requirements.txt
scipy
numpy
pyinstaller
PyQt5
pyqtgraph
pyserial